requirements.txt
wavelink
discord.py
aiohttp
pyyaml
asqlite
uvicorn
fastapi
easy-PIL